requirements.txt
# Core dependencies
prompt-toolkit>=3.0.0
rich>=13.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Market data
yfinance>=0.2.0
pandas-ta>=0.3.14b

# Options pricing
py_vollib>=1.0.1
scipy>=1.10.0

# Database
sqlite3; python_version >= '3.0'

# Utilities
python-dateutil>=2.8.0
tabulate>=0.9.0
